R# rename test classes to match new behaviour
