+++ src/LazyStorage.Tests/PersistenceTests.cs
@@ -22,8 +22,8 @@ namespace LazyStorage.Tests
         {
             _currentStorage = storage;
             var dal = storage.GetStorage();
-            var repo = dal.GetRepository<TestObject>();
-            var obj = new TestObject();
+            var repo = dal.GetRepository<TestObjectIStorable>();
+            var obj = new TestObjectIStorable();
 
             repo.Set(obj);
             dal.Save();
@@ -36,14 +36,14 @@ namespace LazyStorage.Tests
         {
             _currentStorage = storage;
             var dal = storage.GetStorage();
-            var repo = dal.GetRepository<TestObject>();
-            var obj = new TestObject();
+            var repo = dal.GetRepository<TestObjectIStorable>();
+            var obj = new TestObjectIStorable();
 
             repo.Set(obj);
             dal.Save();
 
             var dal2 = storage.GetStorage();
-            var repo2 = dal2.GetRepository<TestObject>();
+            var repo2 = dal2.GetRepository<TestObjectIStorable>();
 
             Assert.True(repo2.Get().Single().ContentEquals(obj), "The object could not be found in the persistent repo");
         }
@@ -54,16 +54,16 @@ namespace LazyStorage.Tests
             _currentStorage = storage;
 
             // Create an object in memory
-            var obj1 = new TestObject();
+            var obj1 = new TestObjectIStorable();
 
             // Insert into the repo
             var dal = storage.GetStorage();
-            var repo = dal.GetRepository<TestObject>();
+            var repo = dal.GetRepository<TestObjectIStorable>();
             repo.Set(obj1);
             dal.Save();
 
             // Make some changes
-            var obj2 = new TestObject
+            var obj2 = new TestObjectIStorable
             {
                 Id = 1,
                 Name = "Test"
@@ -74,7 +74,7 @@ namespace LazyStorage.Tests
             dal.Discard();
 
             var dal2 = storage.GetStorage();
-            var repo2 = dal2.GetRepository<TestObject>();
+            var repo2 = dal2.GetRepository<TestObjectIStorable>();
 
             var testObject = repo2.Get().Single();
             Assert.True(testObject.ContentEquals(obj1), "The object changes were not reverted in the repo");
@@ -89,7 +89,7 @@ namespace LazyStorage.Tests
 
             var repo = storage.GetStorage().GetRepository(converter);
 
-            var obj = new TestObjectNotIStorable
+            var obj = new TestObject
             {
                 Name = "Test",
                 StartDate = DateTime.Now,
@@ -111,7 +111,7 @@ namespace LazyStorage.Tests
 
             var repo = dal.GetRepository(converter);
 
-            var obj = new TestObjectNotIStorable
+            var obj = new TestObject
             {
                 Name = "Test",
                 StartDate = DateTime.Now,
@@ -138,13 +138,13 @@ namespace LazyStorage.Tests
 
             var repo = dal.GetRepository(converter);
 
-            var obj1 = new TestObjectNotIStorable {Name = "Test"};
+            var obj1 = new TestObject {Name = "Test"};
 
             repo.Set(obj1);
             dal.Save();
 
             // Make some changes
-            var obj2 = new TestObjectNotIStorable {Name = "Test"};
+            var obj2 = new TestObject {Name = "Test"};
 
             // Update the object in the repo but discard changes
             repo.Set(obj2);
