+++ src/LazyStorage.Tests/RepositoryWithIStorableObjectTests.cs
@@ -8,19 +8,19 @@ using Xunit;
 
 namespace LazyStorage.Tests
 {
-    public class RepositoryTests
+    public class RepositoryWithIStorableObjectTests
     {
         public static IEnumerable<object[]> Repos => new[]
         {
-            new object[] {new InMemoryRepositoryWithConverter<TestObject>(new StorableConverter<TestObject>())},
-            new object[] {new XmlRepositoryWithConverter<TestObject>("RepositoryTests", new StorableConverter<TestObject>()) },
-            new object[] {new JsonRepositoryWithConverter<TestObject>("RepositoryTests", new StorableConverter<TestObject>()) }
+            new object[] {new InMemoryRepositoryWithConverter<TestObjectIStorable>(new StorableConverter<TestObjectIStorable>())},
+            new object[] {new XmlRepositoryWithConverter<TestObjectIStorable>("RepositoryWithIStorableObjectTests", new StorableConverter<TestObjectIStorable>()) },
+            new object[] {new JsonRepositoryWithConverter<TestObjectIStorable>("RepositoryWithIStorableObjectTests", new StorableConverter<TestObjectIStorable>()) }
         };
 
         [Theory, MemberData(nameof(Repos))]
-        public void CanAddToRepo(IRepository<TestObject> repo)
+        public void CanAddToRepo(IRepository<TestObjectIStorable> repo)
         {
-            var obj = new TestObject();
+            var obj = new TestObjectIStorable();
 
             repo.Set(obj);
             
@@ -30,9 +30,9 @@ namespace LazyStorage.Tests
         }
         
         [Theory, MemberData(nameof(Repos))]
-        public void CanUpdateRepo(IRepository<TestObject> repo)
+        public void CanUpdateRepo(IRepository<TestObjectIStorable> repo)
         {
-            var obj = new TestObject();
+            var obj = new TestObjectIStorable();
 
             repo.Set(obj);
 
@@ -45,9 +45,9 @@ namespace LazyStorage.Tests
         }
 
         [Theory, MemberData(nameof(Repos))]
-        public void CanDelete(IRepository<TestObject> repo)
+        public void CanDelete(IRepository<TestObjectIStorable> repo)
         {
-            var obj = new TestObject();
+            var obj = new TestObjectIStorable();
 
             repo.Set(obj);
             repo.Delete(obj);
@@ -56,10 +56,10 @@ namespace LazyStorage.Tests
         }
 
         [Theory, MemberData(nameof(Repos))]
-        public void CanGetByLinq(IRepository<TestObject> repo)
+        public void CanGetByLinq(IRepository<TestObjectIStorable> repo)
         {
-            var objOne = new TestObject {Id = 1, Name = "one"};
-            var objTwo = new TestObject {Id = 2, Name = "two"};
+            var objOne = new TestObjectIStorable {Id = 1, Name = "one"};
+            var objTwo = new TestObjectIStorable {Id = 2, Name = "two"};
 
             repo.Set(objOne);
             repo.Set(objTwo);
